% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPatrial.R
\name{PlotPatrial}
\alias{PlotPatrial}
\title{To male patrial plot for MFP}
\usage{
PlotPatrial(data, xlim = range(data$PatrialData$x),
  ylim = range(data$PatrialData$pa), log = "", shadow = TRUE,
  shadow.col = rgb(215, 215, 215, maxColorValue = 255), p.pch = 1,
  p.col = "black", p.cex = 1, l.lty = "solid", l.col = "black",
  l.cex = 1, CI.border = NA, CI.col = "grey")
}
\arguments{
\item{data}{results obtained from the RePatrial()}

\item{xlim, ylim}{ranges of x axis and y axis, defaulted as the range of x and y.}

\item{log, }{which axis needs the log scale, defaulted as "", if "x", the x axis scaled; if "xy", both scaled}

\item{shadow, }{logical, whether is 95% sample areas should be maked as a shadowed rectangle}

\item{shadow.col}{the color of the shadow, defaulted as rgb(215,215,215,maxColorValue = 255)}

\item{p.pch}{character of the points, defaulted as 1}

\item{p.col}{color of the points, defaulted as "black"}

\item{p.cex}{size of the points, defaulted as 1}

\item{l.lty}{type of the line, defaulted as "solid"}

\item{l.col}{color of the line, defaulted as "black"}

\item{l.cex}{size of the line, defaulted as 1}

\item{CI.border}{logical, whether the border should be drawn for the CI, default as NA}

\item{CI.col}{the color for the CI band, defaulted as "grey"}
}
\value{
A patrial plot with confidence interval
}
\description{
This function returns with a nice patrial plot with confidence interval using the results from RePatrial()
}
\examples{
x<-runif(100,0,10)
y<-runif(100,0,10)
z=log(x)+0.2*y^2 + rnorm(100,0,2)

library(mfp)
re<-mfp(z~fp(x,4)+fp(y,4))
re0<-RePatrial("y",re)
PlotPatrial(re0,shadow=FALSE)

}
\seealso{
PlotPatrial()
}
